feat(refactor): refactor structural code base of this, to make it cleaner and more maintainable
